+++ integrations/fastify/src/utils/get-javascript-file.ts
